7.1/requirements.txt
streamlit
python-dotenv
openai
chromadb

langchain==0.3.7
langchain-core>=1.0.0,<2.0.0
langchain-community==0.3.5
langchain-openai==0.2.5
langchain-text-splitters==1.0.0
langchain-classic

pypdf
